Set DEFAULT_CONFIG based on currently running kernel (#28)
The path to the DEFAULT_CONFIG depends on the version of the
currently running kernel.  This changes over time (as the default kernel
of Ubuntu gets updated), so typically the path needs to be customised
more or less every time the script is run.  To avoid the need for manual
editing, automatically get the version of the current kernel using
`uname -r`.