Add rt_preempt installation script for Ubuntu 22.04
Updated script to work with the latest kernel version on Ubuntu 22.04.

Some additional changes apart from fixes:
- Download to `~/.cache/rt_preempt_kernel_install` (adheres to
  `$XDG_CACHE_HOME`).  So files do not need to be downloaded again when
  restarting after some issue.
- Change default build directory to `./build` (edit the `BUILD_DIR`
  variable to change it if needed).
- LOCALVERSION is set automatically, so no need to do it manually
  anymore.

Note: The reconfiguration steps in the end are not tested (this was
already done on the machine I used it for, so I commented them out).  I
didn't change anything there, though, so I expect it to work.