Add script to change default option of GRUB
`grub_select_default_kernel.py` is a little script that partially
automates the steps described in [1] to set the default kernel version
in GRUB.

It parses the /boot/grub/grub.cfg file, lists all available options in a
simple curses GUI where the user can select the desired option.
It then updates /etc/default/grub but before actually writing to the
file it shows a diff which has to be confirmed by the user (so if
something goes wrong, the user should hopefully notice it here and
abort).

Disclaimer: The script was written with help of ChatGPT.  I had to fix
some issues and added some stuff but the base it provided was actually
surprisingly good.

[1] https://open-dynamic-robot-initiative.github.io/robot_interfaces/doc/preempt_rt_kernel.html#select-preempt-rt-kernel-by-default